requirements.txt
flask==2.3.3
flask-socketio==5.3.6
openai==1.3.0
python-dotenv==1.0.0
pyaudio==0.2.11
numpy==1.24.3
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1
requests==2.31.0
python-multipart==0.0.6